implementing some phpcs suggestions
