+++ admin/views/download-textbooks.php
@@ -17,8 +17,8 @@ if ( ! defined( 'ABSPATH' ) ) { exit;
 	<h2><?php echo esc_html( get_admin_page_title() ); ?></h2>
 	
 	<?php
-	require( PBT_PLUGIN_DIR . 'includes/modules/catalogue/EquellaFetch.php' );
-	require( PBT_PLUGIN_DIR . 'includes/modules/catalogue/Filter.php' );
+	require( PBT_PLUGIN_DIR . 'inc/modules/catalogue/class-equellafetch.php' );
+	require( PBT_PLUGIN_DIR . 'inc/modules/catalogue/class-filter.php' );
 
 	echo '<h3>Download openly licensed textbooks</h3>';
 
@@ -30,8 +30,8 @@ if ( ! defined( 'ABSPATH' ) ) { exit;
 		echo $textbooks;
 	} else {
 		try {
-			$equellaFetch = new \PBT\Catalogue\EquellaFetch();
-			$filter = new \PBT\Catalogue\Filter( $equellaFetch );
+			$equellaFetch = new \PBT\Modules\Catalogue\EquellaFetch();
+			$filter = new \PBT\Modules\Catalogue\Filter( $equellaFetch );
 			$textbooks = $filter->displayBySubject();
 
 		} catch ( Exception $exc ) {
